requirements.txt
	streamlit
numpy
opencv-python-headless
torch
torchvision
matplotlib
Pillow
